+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @3plate/graph-react
 
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @3plate/graph-core@0.1.18
+
 ## 0.1.18
 
 ### Patch Changes